refactorised into separate class and .h/.cc file
